Added render target

+++ 3D_Engine/3D_Engine/Sphereshaderclass.cpp
@@ -335,7 +335,7 @@ void SphereShaderClass::OutputShaderErrorMessage(ID3D10Blob* errorMessage, HWND 
 	compileErrors = (char*)(errorMessage->GetBufferPointer());
 
 	// Get the length of the message.
-	bufferSize = errorMessage->GetBufferSize();
+	bufferSize = static_cast<unsigned long>(errorMessage->GetBufferSize());
 
 	// Open a file to write the error message to.
 	fout.open("sphere-shader-error.txt");
